NF: extract MigrationContext to separate file
