+++ AnkiDroid/src/main/java/com/ichi2/anki/servicelayer/scopedstorage/MoveDirectory.kt
@@ -20,7 +20,6 @@ import androidx.annotation.VisibleForTesting
 import com.ichi2.anki.model.Directory
 import com.ichi2.anki.servicelayer.scopedstorage.DeleteEmptyDirectory
 import com.ichi2.anki.servicelayer.scopedstorage.MoveDirectoryContent
-import com.ichi2.anki.servicelayer.scopedstorage.migrateuserdata.MigrateUserData.MigrationContext
 import com.ichi2.compat.CompatHelper
 import timber.log.Timber
 import java.io.File
